+++ package/js/swiper.js
@@ -1,5 +1,9 @@
 /**
+<<<<<<< HEAD
  * Swiper 5.3.0
+=======
+ * Swiper 5.3.6
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
  * Most modern mobile touch slider and framework with hardware accelerated transitions
  * http://swiperjs.com
  *
@@ -7,7 +11,11 @@
  *
  * Released under the MIT License
  *
+<<<<<<< HEAD
  * Released on: January 11, 2020
+=======
+ * Released on: February 29, 2020
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
  */
 
 (function (global, factory) {
@@ -1549,10 +1557,20 @@
     }
     // Find slides currently in view
     if (swiper.params.slidesPerView !== 'auto' && swiper.params.slidesPerView > 1) {
+<<<<<<< HEAD
       for (i = 0; i < Math.ceil(swiper.params.slidesPerView); i += 1) {
         var index = swiper.activeIndex + i;
         if (index > swiper.slides.length) { break; }
         activeSlides.push(swiper.slides.eq(index)[0]);
+=======
+      if (swiper.params.centeredSlides) { activeSlides.push.apply(activeSlides, swiper.visibleSlides); }
+      else {
+        for (i = 0; i < Math.ceil(swiper.params.slidesPerView); i += 1) {
+          var index = swiper.activeIndex + i;
+          if (index > swiper.slides.length) { break; }
+          activeSlides.push(swiper.slides.eq(index)[0]);
+        }
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
       }
     } else {
       activeSlides.push(swiper.slides.eq(swiper.activeIndex)[0]);
@@ -1604,7 +1622,11 @@
       var slideProgress = (
         (offsetCenter + (params.centeredSlides ? swiper.minTranslate() : 0)) - slide.swiperSlideOffset
       ) / (slide.swiperSlideSize + params.spaceBetween);
+<<<<<<< HEAD
       if (params.watchSlidesVisibility) {
+=======
+      if (params.watchSlidesVisibility || (params.centeredSlides && params.autoHeight)) {
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
         var slideBefore = -(offsetCenter - slide.swiperSlideOffset);
         var slideAfter = slideBefore + swiper.slidesSizesGrid[i];
         var isVisible = (slideBefore >= 0 && slideBefore < swiper.size - 1)
@@ -1650,7 +1672,11 @@
       isEnd: isEnd,
     });
 
+<<<<<<< HEAD
     if (params.watchSlidesProgress || params.watchSlidesVisibility) { swiper.updateSlidesProgress(translate); }
+=======
+    if (params.watchSlidesProgress || params.watchSlidesVisibility || (params.centeredSlides && params.autoHeight)) { swiper.updateSlidesProgress(translate); }
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
 
     if (isBeginning && !wasBeginning) {
       swiper.emit('reachBeginning toEdge');
@@ -2106,7 +2132,11 @@
 
     var skip = Math.min(swiper.params.slidesPerGroupSkip, slideIndex);
     var snapIndex = skip + Math.floor((slideIndex - skip) / swiper.params.slidesPerGroup);
+<<<<<<< HEAD
     if (snapIndex >= slidesGrid.length) { snapIndex = slidesGrid.length - 1; }
+=======
+    if (snapIndex >= snapGrid.length) { snapIndex = snapGrid.length - 1; }
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
 
     if ((activeIndex || params.initialSlide || 0) === (previousIndex || 0) && runCallbacks) {
       swiper.emit('beforeSlideChangeStart');
@@ -3629,7 +3659,11 @@
     var breakpoint = false;
 
     var points = Object.keys(breakpoints).map(function (point) {
+<<<<<<< HEAD
       if (typeof point === 'string' && point.startsWith('@')) {
+=======
+      if (typeof point === 'string' && point.indexOf('@') === 0) {
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
         var minRatio = parseFloat(point.substr(1));
         var value = win.innerHeight * minRatio;
         return { value: value, point: point };
@@ -4112,7 +4146,11 @@
           startTranslate: undefined,
           allowThresholdMove: undefined,
           // Form elements to match
+<<<<<<< HEAD
           formElements: 'input, select, option, textarea, button, video',
+=======
+          formElements: 'input, select, option, textarea, button, video, label',
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
           // Last click time
           lastClickTime: Utils.now(),
           clickTimeout: undefined,
@@ -6387,9 +6425,15 @@
         gesture.scaleStart = Zoom.getDistanceBetweenTouches(e);
       }
       if (!gesture.$slideEl || !gesture.$slideEl.length) {
+<<<<<<< HEAD
         gesture.$slideEl = $(e.target).closest('.swiper-slide');
         if (gesture.$slideEl.length === 0) { gesture.$slideEl = swiper.slides.eq(swiper.activeIndex); }
         gesture.$imageEl = gesture.$slideEl.find('img, svg, canvas');
+=======
+        gesture.$slideEl = $(e.target).closest(("." + (swiper.params.slideClass)));
+        if (gesture.$slideEl.length === 0) { gesture.$slideEl = swiper.slides.eq(swiper.activeIndex); }
+        gesture.$imageEl = gesture.$slideEl.find('img, svg, canvas, picture, .swiper-zoom-target');
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
         gesture.$imageWrapEl = gesture.$imageEl.parent(("." + (params.containerClass)));
         gesture.maxRatio = gesture.$imageWrapEl.attr('data-swiper-zoom') || params.maxRatio;
         if (gesture.$imageWrapEl.length === 0) {
@@ -6632,8 +6676,13 @@
       var image = zoom.image;
 
       if (!gesture.$slideEl) {
+<<<<<<< HEAD
         gesture.$slideEl = swiper.clickedSlide ? $(swiper.clickedSlide) : swiper.slides.eq(swiper.activeIndex);
         gesture.$imageEl = gesture.$slideEl.find('img, svg, canvas');
+=======
+        gesture.$slideEl = swiper.slides.eq(swiper.activeIndex);
+        gesture.$imageEl = gesture.$slideEl.find('img, svg, canvas, picture, .swiper-zoom-target');
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
         gesture.$imageWrapEl = gesture.$imageEl.parent(("." + (params.containerClass)));
       }
       if (!gesture.$imageEl || gesture.$imageEl.length === 0) { return; }
@@ -6718,8 +6767,13 @@
       var gesture = zoom.gesture;
 
       if (!gesture.$slideEl) {
+<<<<<<< HEAD
         gesture.$slideEl = swiper.clickedSlide ? $(swiper.clickedSlide) : swiper.slides.eq(swiper.activeIndex);
         gesture.$imageEl = gesture.$slideEl.find('img, svg, canvas');
+=======
+        gesture.$slideEl = swiper.slides.eq(swiper.activeIndex);
+        gesture.$imageEl = gesture.$slideEl.find('img, svg, canvas, picture, .swiper-zoom-target');
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
         gesture.$imageWrapEl = gesture.$imageEl.parent(("." + (params.containerClass)));
       }
       if (!gesture.$imageEl || gesture.$imageEl.length === 0) { return; }
@@ -6741,6 +6795,7 @@
       var passiveListener = swiper.touchEvents.start === 'touchstart' && Support.passiveListener && swiper.params.passiveListeners ? { passive: true, capture: false } : false;
       var activeListenerWithCapture = Support.passiveListener ? { passive: false, capture: true } : true;
 
+<<<<<<< HEAD
       // Scale image
       if (Support.gestures) {
         swiper.$wrapperEl.on('gesturestart', '.swiper-slide', zoom.onGestureStart, passiveListener);
@@ -6752,6 +6807,21 @@
         swiper.$wrapperEl.on(swiper.touchEvents.end, '.swiper-slide', zoom.onGestureEnd, passiveListener);
         if (swiper.touchEvents.cancel) {
           swiper.$wrapperEl.on(swiper.touchEvents.cancel, '.swiper-slide', zoom.onGestureEnd, passiveListener);
+=======
+      var slideSelector = "." + (swiper.params.slideClass);
+
+      // Scale image
+      if (Support.gestures) {
+        swiper.$wrapperEl.on('gesturestart', slideSelector, zoom.onGestureStart, passiveListener);
+        swiper.$wrapperEl.on('gesturechange', slideSelector, zoom.onGestureChange, passiveListener);
+        swiper.$wrapperEl.on('gestureend', slideSelector, zoom.onGestureEnd, passiveListener);
+      } else if (swiper.touchEvents.start === 'touchstart') {
+        swiper.$wrapperEl.on(swiper.touchEvents.start, slideSelector, zoom.onGestureStart, passiveListener);
+        swiper.$wrapperEl.on(swiper.touchEvents.move, slideSelector, zoom.onGestureChange, activeListenerWithCapture);
+        swiper.$wrapperEl.on(swiper.touchEvents.end, slideSelector, zoom.onGestureEnd, passiveListener);
+        if (swiper.touchEvents.cancel) {
+          swiper.$wrapperEl.on(swiper.touchEvents.cancel, slideSelector, zoom.onGestureEnd, passiveListener);
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
         }
       }
 
@@ -6768,6 +6838,7 @@
       var passiveListener = swiper.touchEvents.start === 'touchstart' && Support.passiveListener && swiper.params.passiveListeners ? { passive: true, capture: false } : false;
       var activeListenerWithCapture = Support.passiveListener ? { passive: false, capture: true } : true;
 
+<<<<<<< HEAD
       // Scale image
       if (Support.gestures) {
         swiper.$wrapperEl.off('gesturestart', '.swiper-slide', zoom.onGestureStart, passiveListener);
@@ -6779,6 +6850,21 @@
         swiper.$wrapperEl.off(swiper.touchEvents.end, '.swiper-slide', zoom.onGestureEnd, passiveListener);
         if (swiper.touchEvents.cancel) {
           swiper.$wrapperEl.off(swiper.touchEvents.cancel, '.swiper-slide', zoom.onGestureEnd, passiveListener);
+=======
+      var slideSelector = "." + (swiper.params.slideClass);
+
+      // Scale image
+      if (Support.gestures) {
+        swiper.$wrapperEl.off('gesturestart', slideSelector, zoom.onGestureStart, passiveListener);
+        swiper.$wrapperEl.off('gesturechange', slideSelector, zoom.onGestureChange, passiveListener);
+        swiper.$wrapperEl.off('gestureend', slideSelector, zoom.onGestureEnd, passiveListener);
+      } else if (swiper.touchEvents.start === 'touchstart') {
+        swiper.$wrapperEl.off(swiper.touchEvents.start, slideSelector, zoom.onGestureStart, passiveListener);
+        swiper.$wrapperEl.off(swiper.touchEvents.move, slideSelector, zoom.onGestureChange, activeListenerWithCapture);
+        swiper.$wrapperEl.off(swiper.touchEvents.end, slideSelector, zoom.onGestureEnd, passiveListener);
+        if (swiper.touchEvents.cancel) {
+          swiper.$wrapperEl.off(swiper.touchEvents.cancel, slideSelector, zoom.onGestureEnd, passiveListener);
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
         }
       }
 
@@ -6965,6 +7051,12 @@
             }
           }
           swiper.emit('lazyImageReady', $slideEl[0], $imageEl[0]);
+<<<<<<< HEAD
+=======
+          if (swiper.params.autoHeight) {
+            swiper.updateAutoHeight();
+          }
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
         });
 
         swiper.emit('lazyImageLoad', $slideEl[0], $imageEl[0]);
@@ -8320,8 +8412,18 @@
         // var rotateZ = 0
         var translateZ = -translate * Math.abs(offsetMultiplier);
 
+<<<<<<< HEAD
         var translateY = isHorizontal ? 0 : params.stretch * (offsetMultiplier);
         var translateX = isHorizontal ? params.stretch * (offsetMultiplier) : 0;
+=======
+        var stretch = params.stretch;
+        // Allow percentage to make a relative stretch for responsive sliders
+        if (typeof stretch === 'string' && stretch.indexOf('%') !== -1) {
+          stretch = ((parseFloat(params.stretch) / 100) * slideSize);
+        }
+        var translateY = isHorizontal ? 0 : stretch * (offsetMultiplier);
+        var translateX = isHorizontal ? stretch * (offsetMultiplier) : 0;
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
 
         // Fix for ultra small values
         if (Math.abs(translateX) < 0.001) { translateX = 0; }