+++ package/src/components/core/update/updateAutoHeight.js
@@ -10,10 +10,20 @@ export default function (speed) {
   }
   // Find slides currently in view
   if (swiper.params.slidesPerView !== 'auto' && swiper.params.slidesPerView > 1) {
+<<<<<<< HEAD
     for (i = 0; i < Math.ceil(swiper.params.slidesPerView); i += 1) {
       const index = swiper.activeIndex + i;
       if (index > swiper.slides.length) break;
       activeSlides.push(swiper.slides.eq(index)[0]);
+=======
+    if (swiper.params.centeredSlides) activeSlides.push(...swiper.visibleSlides);
+    else {
+      for (i = 0; i < Math.ceil(swiper.params.slidesPerView); i += 1) {
+        const index = swiper.activeIndex + i;
+        if (index > swiper.slides.length) break;
+        activeSlides.push(swiper.slides.eq(index)[0]);
+      }
+>>>>>>> 71efce4731e9ee6ac9544ef3b7bc25c1e52388fe
     }
   } else {
     activeSlides.push(swiper.slides.eq(swiper.activeIndex)[0]);